데이터베이스 연결문제 수정
