+++ resource/template/postdetail.html
@@ -10,15 +10,7 @@
   </head>
   <body>
     <nav class="navigation">
-      <a href="/">홈으로 돌아가기</a>
-      <ul class="postList">
-        <li><a href="?section=성산">성산</a></li>
-        <li><a href="?section=표선">표선</a></li>
-        <li><a href="?section=중문">중문</a></li>
-        <li><a href="?section=안덕">안덕</a></li>
-        <li><a href="?section=한림">한림</a></li>
-        <li><a href="?section=제주시">제주시</a></li>
-      </ul>
+      <a href="/blog">홈으로 돌아가기</a>
       <form action="/:id" method="delete">
         <button id="deleteBtn">삭제</button>
       </form>
