+++ resource/template/write.html
@@ -24,8 +24,10 @@
           <input type="text" id="section" name="section" /><br />
           <label for="name">이름</label>
           <input type="text" id="name" name="name" /><br />
-          <label for="content">추천이유</label>
+          <label for="content">메뉴</label>
           <input type="text" id="content" name="content" /><br />
+          <label for="dec">설명</label><br />
+          <textarea type="text" id="dec" name="dec"></textarea><br />
           <label for="img" class="label-img">사진</label>
           <input type="file" id="img" name="img" /><br />
           <button class="btn-save">저장</button>
